requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
aiohttp==3.8.4
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
altair==4.2.2
    # via streamlit
astroid==2.15.4
    # via pylint
async-timeout==4.0.2
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.12.2
    # via bs4
blinker==1.6.2
    # via streamlit
bs4==0.0.1
    # via -r requirements.in
build==0.10.0
    # via pip-tools
cachetools==5.3.0
    # via streamlit
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   duckduckgo-search
    #   pip-tools
    #   streamlit
dataclasses-json==0.5.7
    # via langchain
decorator==5.1.1
    # via validators
dill==0.3.6
    # via pylint
diskcache==5.6.1
    # via duckduckgo-search
duckduckgo-search==2.9.3
    # via -r requirements.in
entrypoints==0.4
    # via altair
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
func-timeout==4.3.5
    # via -r requirements.in
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via streamlit
importlib-resources==5.12.0
    # via jsonschema
isort==5.12.0
    # via pylint
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
langchain==0.0.161
    # via -r requirements.in
lazy-object-proxy==1.9.0
    # via astroid
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.4
    # via langchain
numpy==1.24.3
    # via
    #   altair
    #   langchain
    #   numexpr
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==0.27.6
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.1
    # via
    #   build
    #   marshmallow
    #   streamlit
pandas==2.0.1
    # via
    #   altair
    #   streamlit
pillow==9.5.0
    # via streamlit
pip-tools==6.13.0
    # via -r requirements.in
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==3.5.0
    # via pylint
protobuf==3.20.3
    # via streamlit
pyarrow==12.0.0
    # via streamlit
pydantic==1.10.7
    # via
    #   langchain
    #   openapi-schema-pydantic
pydeck==0.8.1b0
    # via streamlit
pygments==2.15.1
    # via rich
pylint==2.17.4
    # via -r requirements.in
pympler==1.0.1
    # via streamlit
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via langchain
regex==2023.5.5
    # via tiktoken
requests==2.30.0
    # via
    #   duckduckgo-search
    #   langchain
    #   openai
    #   streamlit
    #   tiktoken
rich==13.3.5
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
soupsieve==2.4.1
    # via beautifulsoup4
sqlalchemy==2.0.12
    # via langchain
streamlit==1.22.0
    # via
    #   -r requirements.in
    #   streamlit-chat
streamlit-chat==0.0.2.2
    # via -r requirements.in
tenacity==8.2.2
    # via
    #   langchain
    #   streamlit
tiktoken==0.3.3
    # via -r requirements.in
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via
    #   build
    #   pylint
    #   pyproject-hooks
tomlkit==0.11.8
    # via pylint
toolz==0.12.0
    # via altair
tornado==6.3.1
    # via streamlit
tqdm==4.65.0
    # via
    #   langchain
    #   openai
typing-extensions==4.5.0
    # via
    #   astroid
    #   pydantic
    #   pylint
    #   rich
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
urllib3==2.0.2
    # via requests
validators==0.20.0
    # via streamlit
wheel==0.40.0
    # via pip-tools
wrapt==1.15.0
    # via astroid
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
